requirements.txt
requests
pandas
seaborn
matplotlib
scipy
python-dotenv
